refactor: FixedVerificationCodeGenerator 위치 변경
