+++ src/test/java/com/gistpetition/api/acceptance/common/TUser.java
@@ -1,5 +1,6 @@
 package com.gistpetition.api.acceptance.common;
 
+import com.gistpetition.api.FixedVerificationCodeGenerator;
 import com.gistpetition.api.user.dto.request.SignInRequest;
 import com.gistpetition.api.user.dto.request.SignUpRequest;
 import com.gistpetition.api.verification.dto.UsernameConfirmationRequest;
